+++ frontend/centero/lib/widgets/admin_call_queue.dart
@@ -1,4 +1,5 @@
-import 'package:centero/services/call_queue.dart';
+import 'dart:async';
+import 'package:centero/services/backend.dart';
 import 'package:flutter/material.dart';
 import 'package:logging/logging.dart';
 
@@ -12,8 +13,19 @@ class AdminCallQueue extends StatefulWidget {
 }
 
 class _AdminCallQueueState extends State<AdminCallQueue> {
+  static const refreshDelay = Duration(seconds: 30);
+
   final Logger logger = Logger('AdminCallQueue');
+
   List<QueuedUserInfo> queue = [];
+  Timer refreshTimer = Timer(refreshDelay, () {});
+
+  @override
+  void initState() {
+    super.initState();
+    refreshTimer = Timer.periodic(refreshDelay, (timer) => fetchLatestQueue());
+    fetchLatestQueue();
+  }
 
   void showPopup(text) {
     ScaffoldMessenger.of(context).showSnackBar(
@@ -22,7 +34,7 @@ class _AdminCallQueueState extends State<AdminCallQueue> {
   }
 
   void fetchLatestQueue() async {
-    final newQueue = await getCallQueue();
+    final newQueue = await Backend.getCallQueue(silent: true);
 
     setState(() {
       if (newQueue == null) {
