+++ scripts/tsc-watch.bat
@@ -1,0 +1,12 @@
+@echo off
+pushd "%~dp0/.."
+
+echo =========================
+echo Compile Backend on Change
+echo =========================
+
+pushd "backend/functions"
+call npm run build:watch
+popd
+
+popd
